fix(lib): Update package.json and index.js exports
This commit refactors the package.json file and updates the exports in
the index.js file. The package.json file now specifies the correct path for the
index.js file in the exports section. Additionally, the index.js file has been
modified to export the library, next, prettier, react, and server modules.

Refactor the package.json and index.js exports

Signed-off-by: Kenton Duprey <kenton@hausofweb.com>